requirements.txt
Flask
python-telegram-bot
image
exifread
universal-analytics-python
emoji
requests
moviepy